+++ src/www/_partials/component-example/_component-example.scss
@@ -46,7 +46,6 @@ $dark-grey: #2b2b2b;
     width: 100%;
     height: 100vh;
     padding-top: 3em;
-    border: 2vh solid get-color(text);
     background-color: $white;
 
     iframe {
